Refactor repository with enhanced concurrency and validation logic
